Move cirq_iqm package to iqm namespace
