+++ examples/demo_valkmusa.py
@@ -18,8 +18,8 @@ the Valkmusa native gateset and connectivity, and then executing it on a simulat
 import cirq
 from demo_common import demo
 
-from cirq_iqm.devices import Valkmusa
-from cirq_iqm.extended_qasm_parser import circuit_from_qasm
+from iqm.cirq_iqm.devices import Valkmusa
+from iqm.cirq_iqm.extended_qasm_parser import circuit_from_qasm
 
 
 def demo_valkmusa(use_qsim: bool = False) -> None:
