+++ pyproject.toml
@@ -22,7 +22,7 @@ dependencies = [
     "numpy",
     "cirq ~= 1.2",
     "ply",  # Required by cirq.contrib.qasm_import
-    "iqm-client >= 13.0, < 14.0"
+    "iqm-client >= 14.0, < 15.0"
 ]
 
 [project.urls]
