+++ src/iqm/cirq_iqm/iqm_operation_mapping.py
@@ -14,7 +14,7 @@
 """Logic for mapping Cirq Operations to the IQM transfer format.
 """
 from cirq.ops import CZPowGate, MeasurementGate, Operation, PhasedXPowGate, XPowGate, YPowGate
-from iqm_client import Instruction
+from iqm.iqm_client import Instruction
 
 
 class OperationNotSupportedError(RuntimeError):
