+++ tests/test_iqm_device.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from cirq_iqm import Adonis, Apollo, Valkmusa
+from iqm.cirq_iqm import Adonis, Apollo, Valkmusa
 
 
 def test_equality_method():
