+++ tox.ini
@@ -30,8 +30,9 @@ commands =
     python --version
     python -m pip --version
     python -m black --check src tests
-    python -m pytest --doctest-modules --isort --mypy --pylint src --verbose --strict
-    python -m pytest --cov --cov-report=term-missing --junitxml=test_report.xml --isort --doctest-modules --mypy --pylint --pylint-rcfile=tests/.pylintrc tests --verbose --strict
+    python -m mypy -p iqm.cirq_iqm
+    python -m pytest --doctest-modules --isort --pylint src --verbose --strict
+    python -m pytest --cov iqm.cirq_iqm --cov-report=term-missing --junitxml=test_report.xml --isort --doctest-modules --mypy --pylint --pylint-rcfile=tests/.pylintrc tests --verbose --strict
 
 [testenv:format]
 description =
